fix: 礼物流水 goods_id 改为 BIGINT，避免超 INTEGER 写入失败
创作者中心 goods_id 会出现 83 万亿量级，INTEGER 写入报 22003，同步被标成失败。

+++ prisma/add_gift_stream.sql
@@ -37,7 +37,7 @@ CREATE TABLE IF NOT EXISTS "received_gift" (
     "uname" TEXT,
     "time" TEXT NOT NULL,
     "ts" BIGINT,
-    "goods_id" INTEGER,
+    "goods_id" BIGINT,
     "gift_id" INTEGER NOT NULL,
     "name" TEXT NOT NULL,
     "num" INTEGER NOT NULL DEFAULT 1,
